EUDEMO read mode

+++ bluemira/codes/fast_ctd/pipeline.py
@@ -16,16 +16,24 @@ from fast_ctd import (
     step_to_brep,
 )
 
-# @dataclass
-# class StepToDagmcPipelineParams:
+from bluemira.base.look_and_feel import bluemira_debug, bluemira_print
 
 
 def step_to_dagmc_pipeline(
     step_file_path: str | Path,
     output_dagmc_model_path: str | Path,
-    comp_name_to_material_map: dict[str, str] | None = None,
+    comp_name_to_material_name_map: dict[str, str] | None = None,
     *,
+    # Optional parameters for fast_ctd
+    minimum_include_volume: float = 1.0,
+    fix_step_to_brep_geometry: bool = False,
+    merge_dist_tolerance: float = 0.001,
+    lin_deflection_tol: float = 0.001,
+    lin_deflection_is_absolute: bool = False,
+    angular_deflection_tol: float = 0.5,
+    run_make_watertight: bool = True,
     enable_debug_logging: bool = False,
+    clean_up: bool = True,
 ):
     """
     Convert a STEP file to a DAGMC model using fast_ctd.
@@ -44,7 +52,7 @@ def step_to_dagmc_pipeline(
         The path to the input STEP file.
     output_dagmc_model_path:
         The path to the output DAGMC model file.
-    comp_name_to_material_map:
+    comp_name_to_material_name_map:
         A dictionary mapping component names to material names.
         If None, the component names are used as the material names.
     enable_debug_logging:
@@ -55,21 +63,24 @@ def step_to_dagmc_pipeline(
 
     intm_materials_csv_file_path = step_file_path.with_suffix(".csv")
     intm_brep_file = step_file_path.with_suffix(".brep")
-    intm_merged_brep_file = step_file_path.with_suffix("_merged.brep")
-    intm_dagmc_file = output_dagmc_model_path.with_suffix("_nwt.h5m")
+    intm_merged_brep_file = step_file_path.with_suffix(".merged.brep")
+    intm_dagmc_file = output_dagmc_model_path.with_suffix(".nwt.h5m")
 
     try:
         comps_info = step_to_brep(
             step_file_path,
             intm_brep_file,
+            minimum_volume=minimum_include_volume,
+            fix_geometry=fix_step_to_brep_geometry,
             enable_logging=enable_debug_logging,
         )
 
         # Map component names to materials and write to the CSV file
-        # If comp_name_to_material_map is None, use the component names as the material
+        # If comp_name_to_material_name_map is None,
+        # use the component names as the material
         mats_list = (
-            [comp_name_to_material_map[e[1]] for e in comps_info]
-            if comp_name_to_material_map
+            [comp_name_to_material_name_map[e[1]] for e in comps_info]
+            if comp_name_to_material_name_map
             else [e[1] for e in comps_info]
         )
         with open(intm_materials_csv_file_path, "w") as f:
@@ -81,22 +92,43 @@ def step_to_dagmc_pipeline(
         merge_brep_geometries(
             intm_brep_file,
             intm_merged_brep_file,
+            dist_tolerance=merge_dist_tolerance,
             enable_logging=enable_debug_logging,
         )
         facet_brep_to_dagmc(
             intm_merged_brep_file,
-            output_h5m_file=output_dagmc_model_path,
+            # Set the output to the output file name
+            # when not running make_watertight
+            output_h5m_file=intm_dagmc_file
+            if run_make_watertight
+            else output_dagmc_model_path,
             materials_csv_file=intm_materials_csv_file_path,
+            lin_deflection_tol=lin_deflection_tol,
+            tol_is_absolute=lin_deflection_is_absolute,
+            ang_deflection_tol=angular_deflection_tol,
             enable_logging=enable_debug_logging,
         )
-        make_watertight(intm_dagmc_file, output_dagmc_model_path)
+        if run_make_watertight:
+            bluemira_debug("Running DAGMC make_watertight...")
+            make_watertight(
+                intm_dagmc_file,
+                output_dagmc_model_path,
+            )
+        else:
+            bluemira_print(
+                "make_watertight not run, run `make_watertight` or "
+                f"`check_watertight {output_dagmc_model_path}` "
+                "manually to make sure the model is watertight."
+            )
     finally:
-        # Clean up intermediate files
-        for file in [
-            intm_materials_csv_file_path,
-            intm_brep_file,
-            intm_merged_brep_file,
-            intm_dagmc_file,
-        ]:
-            if file.exists():
-                file.unlink()
+        if clean_up:
+            # Clean up intermediate files
+            for file in [
+                intm_materials_csv_file_path,
+                intm_brep_file,
+                intm_merged_brep_file,
+                intm_dagmc_file,
+            ]:
+                bluemira_debug(f"Cleaning up: {file}")
+                if file.exists():
+                    file.unlink()
